+++ physical_ai_manager/src/components/TrainingOutputFolderInput.js
@@ -21,7 +21,7 @@ import toast from 'react-hot-toast';
 import { setOutputFolderName, setModelWeightList } from '../features/training/trainingSlice';
 import { useRosServiceCaller } from '../hooks/useRosServiceCaller';
 
-export default function TrainingOutputFolderInput() {
+export default function TrainingOutputFolderInput({ readonly = false }) {
   const dispatch = useDispatch();
 
   const { getModelWeightList } = useRosServiceCaller();
@@ -41,7 +41,8 @@ export default function TrainingOutputFolderInput() {
     'shadow-lg',
     'p-6',
     'w-full',
-    'max-w-md'
+    'max-w-md',
+    'min-w-[250px]'
   );
 
   const classInput = clsx(
@@ -54,8 +55,7 @@ export default function TrainingOutputFolderInput() {
     'focus:outline-none',
     'focus:ring-2',
     'focus:ring-blue-500',
-    'focus:border-transparent',
-    'mb-4'
+    'focus:border-transparent'
   );
 
   const classButton = clsx(
@@ -72,7 +72,10 @@ export default function TrainingOutputFolderInput() {
     'disabled:cursor-not-allowed'
   );
 
-  const classTitle = clsx('text-xl', 'font-bold', 'text-gray-800', 'mb-6', 'text-left');
+  const classTitle = clsx('text-xl', 'font-bold', 'mb-6', 'text-left', {
+    'text-gray-500': readonly,
+    'text-gray-800': !readonly,
+  });
 
   const handleCheckDuplicate = async () => {
     setCheckingDuplicate(true);
@@ -116,7 +119,7 @@ export default function TrainingOutputFolderInput() {
     <div className={classCard}>
       <h1 className={classTitle}>Output folder name</h1>
       <div className="flex flex-row gap-2 items-center justify-center">
-        {!duplicateChecked && <p className="text-gray-500">Please check duplicate</p>}
+        {!duplicateChecked && !readonly && <p className="text-gray-500">Please check duplicate</p>}
         {duplicateChecked &&
           (isOutputFolderAvailable ? (
             <p className="text-blue-500 break-all">{tempOutputFolderName} is available</p>
@@ -127,7 +130,7 @@ export default function TrainingOutputFolderInput() {
       <input
         type="text"
         className={classInput}
-        disabled={checkingDuplicate}
+        disabled={checkingDuplicate || readonly}
         placeholder="Enter your text here..."
         value={tempOutputFolderName}
         onChange={(e) => {
@@ -136,13 +139,20 @@ export default function TrainingOutputFolderInput() {
           setDuplicateChecked(false);
         }}
       />
-      <button
-        className={classButton}
-        onClick={handleCheckDuplicate}
-        disabled={checkingDuplicate || duplicateChecked || tempOutputFolderName === ''}
-      >
-        Check duplicate
-      </button>
+      {!readonly && (
+        <>
+          <div className="mb-4" />
+          <button
+            className={classButton}
+            onClick={handleCheckDuplicate}
+            disabled={
+              checkingDuplicate || duplicateChecked || tempOutputFolderName === '' || readonly
+            }
+          >
+            Check duplicate
+          </button>
+        </>
+      )}
     </div>
   );
 }